updated categorisation entrypoint, removed "example" entrypoint
