+++ littleTalkApp/templates/exercises/categorisation.html
@@ -16,6 +16,6 @@
     {% vite_react_refresh %}
     
     <!-- Exercise entry point -->
-    {% vite_asset 'frontend/src/entrypoints/categorisation-example.tsx' %}
+    {% vite_asset 'frontend/src/entrypoints/categorisation.tsx' %}
 </body>
 </html>
